refactor: split monolithic main.go into domain-specific command files
Extract 2000+ line main.go into 13 focused files with DI via App struct:
- app.go: App struct, NewApp(), Manager(), validators, initLogging()
- cmd_container.go: init/create/start/stop/lock/unlock/logs/delete/list/status
- cmd_shell.go: ssh/shell/exec (incus exec path for bladerunner)
- cmd_config.go: config/env display
- cmd_theme.go: interactive theme picker and previews
- cmd_snapshot.go: snapshot CRUD
- cmd_mount.go: mount add/remove/list with seatbelt auth
- cmd_vm.go: VM lifecycle commands
- cmd_doctor.go: health checks
- cmd_image.go: image build/list/exists/publish/lineage
- cmd_state.go: state history/show
- cmd_help.go: help/version with dashboard
- main.go: thin 91-line dispatch

Eliminates appConfig global; all state flows through App struct.
Fix all errcheck lint issues across bladerunner, doctor, and cmd files.
